+++ src/main/java/be/ugent/rml/cli/Main.java
@@ -2,6 +2,7 @@ package be.ugent.rml.cli;
 
 import be.ugent.rml.*;
 import be.ugent.rml.functions.FunctionLoader;
+import be.ugent.rml.functions.lib.GrelProcessor;
 import be.ugent.rml.records.RecordsFactory;
 import be.ugent.rml.store.QuadStore;
 import be.ugent.rml.store.SimpleQuadStore;
@@ -15,7 +16,12 @@ import org.slf4j.LoggerFactory;
 
 import java.io.*;
 import java.time.Instant;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Properties;
+import java.util.Map;
+import java.util.HashMap;
 
 public class Main {
 
@@ -145,8 +151,9 @@ public class Main {
                 if (fOptionValue == null) {
                     executor = new Executor(rmlStore, factory);
                 } else {
-                    File functionFile = Utils.getFile(fOptionValue);
-                    FunctionLoader functionLoader = new FunctionLoader(functionFile);
+                    Map<String, Class> libraryMap = new HashMap<>();
+                    libraryMap.put("GrelFunctions", GrelProcessor.class);
+                    FunctionLoader functionLoader = new FunctionLoader(null, null, libraryMap);
                     executor = new Executor(rmlStore, factory, functionLoader);
                 }